+++ Source/controls/plrctrls.cpp
@@ -2036,7 +2036,7 @@ void PerformSecondaryAction()
 	}
 }
 
-void QuickCast(int slot)
+void QuickCast(size_t slot)
 {
 	auto &myPlayer = Players[MyPlayerId];
